+++ Lv2_ps.py
@@ -61,7 +61,7 @@ def whole(eventfile,par_list,tbin_size,mode,ps_type,oversampling,xlims,vlines):
     counts = np.ones(len(times))
 
     shifted_t = times-times[0]
-    t_bins = np.linspace(0,np.ceil(shifted_t[-1]),np.ceil(shifted_t[-1])*1/tbin_size+1)
+    t_bins = np.linspace(0,np.ceil(shifted_t[-1]),int(np.ceil(shifted_t[-1])*1/tbin_size+1))
     summed_data, bin_edges, binnumber = stats.binned_statistic(shifted_t,counts,statistic='sum',bins=t_bins) #binning the time values in the data
 
     event_header = fits.open(eventfile)[1].header
